Make specs headers more independent
Changing, e.g., argv_helper.h should not result in the recompilation
of all specs, although only two of the specs use the argv_helper.
Hence, the inclusion of argv_helper.h is done where it is used, and
not generally in specs.h

This is also done for the fake headers.

While at it, the specs headers are made self-sufficient.

The collecting headers fakes.h and util.h are removed.