Remove TODO to print unknown options
It could be helpful for the user to see when she used options
that are unknown. However, bandit's usage is simple enough to
assume that the usage screen is helpful enough.

Also printing unknown options may be confusing if user code adds
custom options (which is a hacky thing, so far...)
Hence, this commit removes the TODO marker.

As soon as there is a nice way to add custom options to bandit (see #92),
one should add such output.