Replace skip/only patterns by filter chain
A filter chain is now used by run policies instead of
skip/only patterns.

This allows a more intuitive usage of (multiple) --skip
and --only options.

However, this comes with a compatibility-breaking change:
Up to this commit, context skips were ignored if there are
matching "only"s. This behavior is now changed. --skip
and --only apply a filter (the one or the other way),
no matter in what "stage" they match.

This resolves #88.