specs: Shorten filenames and remove .spec suffix
It is clear that all source code files in specs/
are specs, so we do not have to mention this in the
filename.

Also having more than one dot in a filename might
lead to problems on some systems.

While at it, many filenames are shortened and the
colorizer.cpp is moved out of reporters/ because it
is not testing a reporter.