index.md: Replace badges by shields.io badges
This results in a SPOF now for the badges but also less tracking for the users.
The colors changed a little and the bade label texts on the left-hand side are
now more expressive in the meaning that the semantics are made clear without
advertising brand names. The AppVeyor logo is also dropped from the AppVeyor badge.