Add BUILD_TYPE=coverity to Travis CI config
We want that Travis CI automatically submits information to Coverity Scan
whenever code is pushed to the permanent coverity-scan branch.

The purpose of the coverity-scan branch is two-fold: usually, it should
match the latest commit in master or one of its ancestors, that is, it should
not diverge from master. The only reason to diverge from master is to fix
defects found by Coverity Scan. In this case, the coverity-scan branch will
be merged into master afterwards, so that two branches do not diverge.

Because of this purpose, the Travis CI scripts are changed as follows:
when we have pushed to the coverity-scan branch, run the Coverity Scan
script (see travis-ci/travis-ci#1975) and do nothing else;
when we push to another branch: business as usual but do NOT run the
Coverity Scan script.

This behavior is implemented by simply adding BUILD_TYPE=coverity
and some trivial conditional premature exiting in travis_common.sh