Copy conditional-branch logic into .travis.yml
Travis CI is slow at the moment. This showed that it is useful
to exclude whole build matrix entries on certain branches instead
on only letting the script do nothing on those branches (but the
environment/image still has to be spawned and initialized which
takes a lot of time).

The chosen solution is the one mentioned in travis-ci/travis-ci#2778

We kept the logic in the shell scripts though.