Allow use of Snowhouse versions older than 5.0.0
Although we currently include Snowhouse as a submodule,
a goal is to make Bandit independent of a specific assertion framework.
The idea is that we use Snowhouse to test bandit itself, but any
Bandit user might decide for another assertion framework.

A first step in this direction is to make the Snowhouse adapter work
with older version of Snowhouse, by using the C++ preprocessor and
checking for SNOWHOUSE_MAJOR.