[WIP] Introduce linebreak before Snowhouse messages
The failure formatters of Bandit are intended for
single-line failure messages.
Multi-line failure messages, as Snowhouse produces them,
are misaligned and hard to read.

The solution is to insert a linebreak before the message.
However, in contrast to PR #39 we do not add it for the
failure formatter, but for the Snowhouse adapter, because
I believe the issue is Snowhouse-specific (and specific to
every other multi-line failure message producer).

Closes #39.

Note that this is a breaking change.

TODO:
 - change doc
 - fix specs